Add files via upload
Instead of banning leaping riders if they might be able to attack directly, the code now checks if an given leaping rider can and rejects the board if it can.

New Feature: If the board is quite large, there is a chance that some spaces in the middle will have special properties, adding an aspect of terrain to the battle. The terrain is mirrored for both sides of the battlefield.

New Feature: There is now a chance that the board will mirror the second players pieces, instead of copying them directly.